All only css files renamed to styles.jsx and connections are updated.
