+++ src/components/ErrorModal.jsx
@@ -1,5 +1,5 @@
-import { ErrorCard } from './cards/Cards';
-import { SmallButton } from './buttons/Button';
+import { ErrorCard } from './cards/Cards.styles';
+import { SmallButton } from './buttons/Button.styles';
 import GhostIcon from './icons/Ghost';
 
 export function ErrorModal({ errorMessage, onClose }) {
