+++ src/components/GhostContainer.styles.jsx
