+++ src/pages/Questions.jsx
@@ -2,11 +2,11 @@ import { useQuestions } from "../hooks/useQuestions";
 import { useNavigate, useParams } from "react-router-dom";
 
 // Components
-import { HeadingCard } from "../components/cards/Cards";
-import { ButtonRow } from "../components/buttons/ButtonRow";
-import { SmallButton, SmallLightButton } from "../components/buttons/Button";
-import { QuestionCard } from "../components/cards/Cards";
-import { NewQuestionButton } from "../components/buttons/Button";
+import { HeadingCard } from "../components/cards/Cards.styles";
+import { ButtonRow } from "../components/buttons/ButtonRow.styles";
+import { SmallButton, SmallLightButton } from "../components/buttons/Button.styles";
+import { QuestionCard } from "../components/cards/Cards.styles";
+import { NewQuestionButton } from "../components/buttons/Button.styles";
 
 // Language
 import { useLanguage } from "../hooks/useLanguage";
